+++ wpcleanadmin/includes/class-wpca-dashboard.php
@@ -1,4 +1,4 @@
-﻿<?php
+<?php
 /**
  * WPCleanAdmin Dashboard Class
  *