+++ wpcleanadmin/assets/js/wpca-login.js
@@ -2,7 +2,7 @@
  * WP Clean Admin Login JavaScript
  *
  * @package WPCA
- * @version 1.8.0
+ * @version 1.8.2
  * @author Sut
  * @author URI: https://github.com/sutchan
  * @since 1.7.15