+++ wpcleanadmin/includes/modules/admin/settings/class-wpca-settings-scripts.php
@@ -3,7 +3,7 @@
  * WPCleanAdmin Settings Scripts Class
  *
  * @package WPCleanAdmin\Modules\Admin\Settings
- * @version 1.8.0
+ * @version 1.8.2
  * @author Sut
  * @author URI: https://github.com/sutchan
  * @since 1.7.15