+++ wpcleanadmin/includes/settings/fields/class-wpca-security-settings-fields.php
@@ -3,7 +3,7 @@
  * WPCleanAdmin Security Settings Fields Class
  *
  * @package WPCleanAdmin
- * @version 1.8.0
+ * @version 1.8.2
  * @update_date 2026-01-28
  * @author Sut
  * @author URI: https://github.com/sutchan